backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
ezdxf
matplotlib
Pillow
python-dotenv
